refactor(core): move GlobalErrorHandler to /logger
