+++ packages/logger/global-error-handler/ng-package.json
@@ -1,0 +1,5 @@
+{
+  "lib": {
+    "entryFile": "./index.ts"
+  }
+}
