+++ tsconfig.json
@@ -96,6 +96,9 @@
       "@xm-ngx/logger/testing": [
         "packages/logger/testing"
       ],
+      "@xm-ngx/logger/global-error-handler": [
+          "packages/logger/global-error-handler"
+      ],
       "@xm-ngx/timeline": [
         "packages/timeline"
       ],
